requirements.txt
feedparser>=6.0,<7.0
pyyaml>=6.0,<7.0
requests>=2.31,<3.0
openai>=1.0,<2.0
anthropic>=0.30,<1.0
beautifulsoup4>=4.12,<5.0
gspread>=6.0,<7.0
google-auth>=2.0,<3.0
google-genai>=1.0,<2.0
